fix divide by zero bugs

+++ riscv/insns/divw.h
@@ -1,5 +1,5 @@
 require_xpr64;
-if(RS2 == 0)
+if(int32_t(RS2) == 0)
   RD = UINT64_MAX;
 else if(int32_t(RS1) == INT32_MIN && int32_t(RS2) == -1)
   RD = RS1;
